Move ensureColorScheme to state and fix minor color related bugs (#3649)
* fix bugs and move ensureColorScheme to fiftyone/state

* consolidate fallback color scheme handling

* lint

---------
