+++ app/packages/app/src/useEvents/useSetColorScheme.ts
@@ -1,7 +1,6 @@
-import { useSessionSetter } from "@fiftyone/state";
+import { ensureColorScheme, useSessionSetter } from "@fiftyone/state";
 import { useCallback } from "react";
 import { EventHandlerHook } from "./registerEvent";
-import { ensureColorScheme } from "./utils";
 
 const useSetColorScheme: EventHandlerHook = () => {
   const setter = useSessionSetter();
