+++ app/packages/app/src/useEvents/utils.ts
@@ -1,5 +1,5 @@
 import { ColorSchemeInput } from "@fiftyone/relay";
-import { State, useSessionSetter } from "@fiftyone/state";
+import { State, ensureColorScheme, useSessionSetter } from "@fiftyone/state";
 import { toCamelCase } from "@fiftyone/utilities";
 import { atom } from "recoil";
 import { AppReadyState } from "./registerEvent";
@@ -9,34 +9,13 @@ export const appReadyState = atom<AppReadyState>({
   default: AppReadyState.CONNECTING,
 });
 
-export const ensureColorScheme = (
-  colorScheme: any,
-  defaultValue: any
-): ColorSchemeInput => {
-  colorScheme = toCamelCase(colorScheme);
-  return {
-    colorPool: colorScheme.colorPool || defaultValue.colorPool,
-    colorBy: colorScheme.colorBy || defaultValue.colorBy,
-    fields: colorScheme.fields as ColorSchemeInput["fields"],
-    multicolorKeypoints:
-      colorScheme.multicolorKeypoints || colorScheme.multicolorKeypoints,
-    opacity: colorScheme.opacity || defaultValue.opacity,
-    showSkeletons: colorScheme.showSkeletons == false ? false : true,
-  };
-};
-
 export const processState = (
   setter: ReturnType<typeof useSessionSetter>,
   state: any
 ) => {
   setter(
     "colorScheme",
-    ensureColorScheme(state.color_scheme as ColorSchemeInput, {
-      colorPool: [],
-      colorBy: "field",
-      opacity: 0.7,
-      multicolorKeypoints: false,
-    })
+    ensureColorScheme(state.color_scheme as ColorSchemeInput)
   );
   setter("sessionGroupSlice", state.group_slice);
   setter("selectedSamples", new Set(state.selected));
